+++ src/dither/mod.ts
@@ -1,0 +1,3 @@
+export * from "./fs.ts";
+export * from "./quick.ts";
+export * from "./sierra.ts";
