+++ src/mod.ts
@@ -1,20 +1,12 @@
-export { getProminentColor, reducePalette, quantizeByMedianCut, quantizeByPopularity } from "./quantize.ts";
+export * from "./quantize/mod.ts";
 
 export { ColorHistogram, getHistogram } from "./histogram.ts";
 
 export { getPixels } from "./pixels.ts";
 
-export {
-  floydSteinbergDither,
-  midPointDither,
-  monochromeDither,
-  noDither,
-  quickDither,
-  twoRowSierraDither,
-  quickTwoRowDither,
-} from "./dither.ts";
+export * from "./dither/mod.ts";
 
-export { Image } from "./image.ts"
+export { Image } from "./structures/image.ts";
 
 // For backwards compat. TODO: Remove soon
 export { Color } from "../deps/color.ts";
