+++ src/quantize/mod.ts
@@ -1,0 +1,3 @@
+export { quantizeBymedianCut } from "./medianCut.ts";
+
+export { quantizeByPopularity } from "./popularity.ts";
